Pop-up confirmation added
Added a basic popup-menu with yes/no answers.
