--- Assets/Scripts/Managers/ButtonManager.cs
@@ -1,24 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class ButtonManager : MonoBehaviour {
-
-    public void NewGame(string newGameLevel)
-    {
-        SceneManager.LoadScene(newGameLevel);
-    }
-
-    public void ExitGame()
-    {
-        Application.Quit();           
-    }
-
-    void Start()
-    {
-        Cursor.visible = true;
-    }
-
-
-}
