+++ Assets/Scripts/Managers/MenuManager.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 93743ba914ff0784a80e7cdeedd5c286
-timeCreated: 1488118457
+guid: 6d29302b53b20184f8ec28d4df25f9ee
+timeCreated: 1488563367
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
