--- Assets/Videos.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: b4d2117da7235da43ba94026ae1b7916
-folderAsset: yes
-timeCreated: 1488056588
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
