+++ Library/LastSceneManagerSetup.txt
@@ -1,4 +1,4 @@
 sceneSetups:
-- path: Assets/Scenes/Level 01.unity
+- path: Assets/Scenes/Menu.unity
   isLoaded: 1
   isActive: 1
