+++ Survival-Shooter.csproj
@@ -115,9 +115,9 @@
     <Compile Include="Assets\Scripts\Fanton\FantonMovement.cs" />
     <Compile Include="Assets\Scripts\Lovisa\LovisaMovement.cs" />
     <Compile Include="Assets\Scripts\Lovisa\LovisaPunching.cs" />
-    <Compile Include="Assets\Scripts\Managers\ButtonManager.cs" />
     <Compile Include="Assets\Scripts\Managers\EnemyManager.cs" />
     <Compile Include="Assets\Scripts\Managers\GameOverManager.cs" />
+    <Compile Include="Assets\Scripts\Managers\MenuManager.cs" />
     <Compile Include="Assets\Scripts\Managers\RageManager.cs" />
     <Compile Include="Assets\Scripts\Managers\ScoreManager.cs" />
     <Compile Include="Assets\Scripts\Managers\VideoManager.cs" />
